feat(QiunetUtils): 定义些常用的functional 和 collector

+++ QiunetUtils/src/main/java/org/qiunet/utils/monitor/DefaultMonitor.java
@@ -67,7 +67,7 @@ public class DefaultMonitor<Type, SubType> implements IMonitor<Type, SubType> {
 						  IMonitorTrigger<Type, SubType> trigger,
 						  final int triggerTime,
 						  final TimeUnit unit) {
-		this(numMapping, trigger, subType -> (int) unit.toMillis(triggerTime));
+		this(numMapping, trigger, subType -> unit.toMillis(triggerTime));
 	}
 
 	public DefaultMonitor(IMonitorTriggerNumMapping<SubType> numMapping,
